+++ CHANGELOG.md
@@ -11,10 +11,13 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds Storybook interaction tests for `CheckboxGroup`, `DatePicker`, and `Slider` components.
+- Adds `closeOnBlur` prop to `MultiSelect` component which, when set to true, closes the component if it loses focus.
 
 ### Updates
 
 - Updates base font stack to only `system-ui` and `sans-serif`.
+- Updates `Accordion` to close focused panel when "esc" key is pressed.
+- Updates `Accordion` to close panel when element within panel is focused and "esc" key is pressed.
 
 ## 3.1.6 (June 20, 2024)
 
@@ -54,9 +57,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Exports `useMediaQuery` and `chakra` from Chakra UI.
-- Updates `Accordion` to close focused panel when "esc" key is pressed.
 - Updates the `TextInput` and `SearchBar` components to better associate the input element to the entire component's helper text.
-- Updates `Accordion` to close panel when element within panel is focused and "esc" key is pressed.
 
 ## 3.1.4 (May 23, 2024)